releaseAutomat: release 4.5.0
It contains Bootstrap 5.3.2, Node,je version bump to 18.x,
security updates and major package upgrades.

* Node.js version bump to 18.x
* Update GitHub Actions to latest versions

* Update Bootstrap to 5.3.2

* fix: `gulp-todo.js`
* fix: callbacks
* refactor: logging
* refactor: documentation
* refactor: linters

+++ gulp-tasks/gulp-todo.js
@@ -7,14 +7,21 @@ const through2 = require('through2');
 const todo = require('gulp-todo');
 
 /**
- * @description A function to generate a TODO.md file
- * @param {object} params - The function parameters.
- * @returns {Stream} Compiled file
+ * Builds a TODO.md file by scanning JavaScript, CSS, SCSS, and Markdown files for TODO comments.
+ * @param {object} params - Optional parameters for the buildTodo function.
+ * @param {Function} params.cb - Callback function to be executed after the TODO file is created.
+ * @param {boolean} params.verbose - Flag indicating whether to log verbose output.
+ * @returns {void} - Gulp stream that generates the TODO.md file.
+ * @throws {Error} - If the callback parameter is not a function.
  */
 const buildTodo = (params = {}) => {
   let todoExist = false;
-
   const filePath = './TODO.md';
+  const cb = params.cb || (() => {});
+
+  if (typeof cb !== 'function') {
+    throw new Error('Callback in params should be of type function.');
+  }
 
   if (fs.existsSync(filePath)) {
     fs.unlinkSync(filePath);
@@ -26,33 +33,33 @@ const buildTodo = (params = {}) => {
     .pipe(todo())
     .pipe(replace('### ', '# '))
     .pipe(
-      through2.obj(function processFile(file, _, cb) {
+      through2.obj(function processFile(file, _, callback) {
         if (file.todos?.length) {
           todoExist = true;
           this.push(file);
         }
-        cb();
-      })
+        callback();
+      }),
     )
     .pipe(gulp.dest('./'))
     .on('end', async () => {
       if (!todoExist) {
         log('         No TODOs found.');
-        params.cb?.();
+        cb();
         return;
       }
 
       try {
         await exec(
-          `npx remark-cli -q ${filePath} -o -- && git add ${filePath}`
+          `npx remark-cli -q ${filePath} -o -- && git add ${filePath}`,
         );
         if (params.verbose) {
           log('         ToDos created.');
         }
       } catch (error) {
-        console.error(`exec error: ${error}`);
+        log.error(`exec error: ${error}`);
       }
-      params.cb?.();
+      cb();
     });
 };
